style: remove Advanced view and group categories under General and Features folders
